Move the plugin to chat-plugins
