+++ chat-plugins/github.js
